move RightPaneStore inside RightPane folder
