+++ src/components/RightPane/CoursePane/SearchForm/DeptSearchBar/MobileDeptSelector.js
@@ -5,7 +5,7 @@ import FormControl from '@material-ui/core/FormControl';
 import Select from '@material-ui/core/Select';
 import depts from './depts';
 import { updateFormValue } from '../../../../stores/RightPaneStore';
-import RightPaneStore from '../../../stores/RightPaneStore.js';
+import RightPaneStore from '../../../RightPaneStore';
 import { withStyles } from '@material-ui/core/styles';
 
 const style = {
