+++ src/components/RightPane/CoursePane/SearchForm/TermSelector.js
@@ -3,7 +3,7 @@ import InputLabel from '@material-ui/core/InputLabel';
 import MenuItem from '@material-ui/core/MenuItem';
 import FormControl from '@material-ui/core/FormControl';
 import Select from '@material-ui/core/Select';
-import RightPaneStore from '../../../../stores/RightPaneStore.js';
+import RightPaneStore from '../../RightPaneStore';
 import { termData } from '../../../../termData';
 
 class TermSelector extends PureComponent {
