+++ src/components/RightPane/RightPaneRoot.js
@@ -3,7 +3,7 @@ import { Tab, Tabs, Typography, Paper } from '@material-ui/core';
 import { FormatListBulleted, MyLocation, Search } from '@material-ui/icons';
 import AddedCoursePane from './AddedCourses/AddedCoursePane';
 import CoursePane from './CoursePane/CoursePaneRoot';
-import RightPaneStore from '../../stores/RightPaneStore';
+import RightPaneStore from './RightPaneStore';
 import { isDarkMode } from '../../helpers';
 import darkModeLoadingGif from './CoursePane/SearchForm/Gifs/dark-loading.gif';
 import loadingGif from './CoursePane/SearchForm/Gifs/loading.gif';
